fix: use background instead of foreground for `Error` group (#98)
I noticed that all builtin colorschemes use bg instead of fg for this group,
and some plugins like mini.trailspace rely on that. The only other usage
of this group that I found is for syntax errors, and in my opinion, with
bg it is more noticeable than with fg.

It's not that important, but since the default colorschemes use bg, I
think it's safer to do the same.